pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "abet-syllabus"
version = "0.1.0"
description = "CLI tool to generate ABET-compliant course syllabi from university course specifications"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["abet", "syllabus", "accreditation", "education"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
]

dependencies = [
    "pdfplumber>=0.10",
    "python-docx>=1.0",
    "PyYAML>=6.0",
    "defusedxml>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
abet-syllabus = "abet_syllabus.cli:main_cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
